requirements.txt
streamlit
folium
streamlit_folium
geopandas
shapely
transformers
Pillow
torch
rasterio
numpy
pandapower
matplotlib